feat: adding hash params for modals in manage
