+++ src/routes/(docs)/+layout.svelte
@@ -85,7 +85,7 @@
             <!-- Document Icon - Replace with your own logo -->
             <img src="https://kener.ing/logo.png" class="h-8 w-8" alt="" />
             <span class="text-xl font-medium">Kener Documentation</span>
-            <span class="me-2 rounded border px-2.5 py-0.5 text-xs font-medium"> 3.0.11 </span>
+            <span class="me-2 rounded border px-2.5 py-0.5 text-xs font-medium"> 3.0.12 </span>
           </a>
         </div>
